Allow RCWT file creation to work when providing muxed files
The original RCWT support was baked into the routine which parses
the old VANC file format.  Allow it to work when parsing muxed files
as well.

Note that this also introduces support for embeddeding the
relative time offset into the RCWT file, which is technically
required but wasn't being done with the legacy VANC format (since
there was no timing information available to embed).