Fix case where changes in format not reflected in mux files
If the SDI source format changes, we reconfigure the input but we
were still writing the original value into the mux file.  This would
typically result in the mux file containing "NTSC" (which is what the
capture utility runs at by default) through the entire output mux
file.

Make sure the values written to the mux file reflect the actual
format at the time the frame is captured.