Cleanup the correct mux object
Because of cut/paste code from capture.cpp, we were trying to
close the "muxedSession" variable, which is never used in the
project.  Because the pointer was NULL the close was never called,
and hence it wasn't hurting anything.  But we also weren't
actually trying to close the real mux session.

Get rid of the unused variable, and close the correct mux object.