Refactor the way VANC is inserted into individual frames
In the existing code, we would create the ancillary region once
and tie it to a frame that gets reused over and over.  This
approach causes issues in that duplicate VANC messages may
appear across multiple frames when the frame gets rescheduled.

Explicitly create a new ancillary region on each scheduled frame,
and tie it to that frame when scheduled.