Separate out burnin counters from fillVideo()
The original fillVideo() routine filled the video region with
colorbars as well as burning in counters if desired.  However this
prevents the burn-in functionality from working if the video came
from a mux file.

Move the routine which burns in the counter (and will in the future
perform other operations) into a separate routine that gets called
regardless of whether the source video came from a mux file or was
generated internally.