Don't assume mux files will always contain video.
The existing code would assume that we would only generate colorbars
if the user wasn't providing a mux file, and if a mux file was
provided it would always contain video.

Eliminate these assumptions, such that if a mux file doesn't
contain video then we fill the video with colorbars before
dispatching the frame to the hardware output.

This problem manifested when providing a mux file that only
contained VANC data (created using klvanc_capture with "-ev").
The visual output would just cycle between the three frames we
pre-rolled with colorbars (i.e. frames with count 1,2,3,1,2,3...).